Fix/return user profile on create user
* fix: changed create user endpoint to return UserProfileDto

+++ src/main/java/com/itesm/domain/models/Role.java
@@ -9,10 +9,10 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class Role {
-    private Integer id;
+    private Byte id;
     private String name;
 
-    public Role(Integer id) {
+    public Role(Byte id) {
         this.id = id;
     }
 }
